new camera movement; multithreading rendering(unstable maybe)

+++ funcs.c
@@ -6,37 +6,37 @@
 /*   By: zjeyne-l <zjeyne-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/16 18:45:21 by zjeyne-l          #+#    #+#             */
-/*   Updated: 2019/10/24 13:10:54 by zjeyne-l         ###   ########.fr       */
+/*   Updated: 2019/10/28 23:57:07 by zjeyne-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "rt.h"
 
-float		ft_max(float a, float b)
+double		ft_max(double a, double b)
 {
 	if (a > b)
 		return (a);
 	return (b);
 }
 
-float       ft_dot_prod(t_vec3 *a, t_vec3 *b)
+double       ft_dot_prod(t_vec3 *a, t_vec3 *b)
 {
-    float dot = a->x * b->x + a->y * b->y + a->z * b->z;
+    double dot = a->x * b->x + a->y * b->y + a->z * b->z;
     return (dot);
 }
 
-float		ft_vec_len(t_vec3 *vec)
+double		ft_vec_len(t_vec3 *vec)
 {
-	float len = sqrtf(vec->x * vec->x + vec->y * vec->y + vec->z * vec->z);
+	double len = sqrtf(vec->x * vec->x + vec->y * vec->y + vec->z * vec->z);
 	return (len);
 }
 
 t_vec3		*ft_vec_normalize(t_vec3 *vec)
 {
-	float len = ft_vec_len(vec);
-	vec->x = (float)vec->x / (float)len;
-	vec->y = (float)vec->y / (float)len;
-	vec->z = (float)vec->z / (float)len;
+	double len = ft_vec_len(vec);
+	vec->x = (double)vec->x / (double)len;
+	vec->y = (double)vec->y / (double)len;
+	vec->z = (double)vec->z / (double)len;
 	return (vec);
 }
 
@@ -49,7 +49,7 @@ int			ft_color_convert(int color, double lum)
 	return (((r & 0xFF) << 16) + ((g & 0xFF) << 8) + ((b & 0xFF)));
 }
 
-int         ft_sum_color(int c1, int c2, float k1, float k2)
+int         ft_sum_color(int c1, int c2, double k1, double k2)
 {
     int r1 = (c1 >> 16) & 0xFF;
     int g1 = (c1 >> 8) & 0xFF;
