+++ normal.c
@@ -6,7 +6,7 @@
 /*   By: zjeyne-l <zjeyne-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/23 17:41:14 by zjeyne-l          #+#    #+#             */
-/*   Updated: 2019/10/28 01:18:55 by zjeyne-l         ###   ########.fr       */
+/*   Updated: 2019/10/28 23:57:07 by zjeyne-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ t_vec3		*ft_cone_normal_calc(t_vec3 *normal, t_vec3 *dir, t_vec3 *point, t_obj *
     normal->x = point->x - obj->c->x;
     normal->y = point->y - obj->c->y;
     normal->z = point->z - obj->c->z;
-    float side_len = sqrt(ft_dot_prod(normal, normal));
+    double side_len = sqrt(ft_dot_prod(normal, normal));
     side_len /= cos(obj->radius);
     if (ft_dot_prod(normal, obj->normal) < 0)
         side_len = -side_len;
@@ -62,7 +62,7 @@ t_vec3		*ft_cylinder_normal_calc(t_vec3 *normal, t_vec3 *dir, t_vec3 *point, t_o
     normal->x = point->x - obj->c->x;
     normal->y = point->y - obj->c->y;
     normal->z = point->z - obj->c->z;
-    float dis = ft_dot_prod(normal, obj->normal);
+    double dis = ft_dot_prod(normal, obj->normal);
     normal->x = obj->c->x + obj->normal->x * dis;
     normal->y = obj->c->y + obj->normal->y * dis;
     normal->z = obj->c->z + obj->normal->z * dis;
